Blood management system
This project was developed by me and my friend Rupesh during our free time. We learned the necessary skills from YouTube tutorials, and it was a great learning experience for both of us.

We created this Blood Management System after observing the current scenario in Nepal, where people often struggle to find blood during emergencies. Our goal was to build a system that makes it easier to connect donors with those in need, helping to save lives more efficiently.